LLD-2/class-14-code-book-my-show/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-filter==24.3
